+++ .github/workflows/publish-snapshot.yml
@@ -28,3 +28,4 @@ jobs:
           MAVEN_SPACE_USERNAME: ${{ secrets.MAVEN_SPACE_USERNAME }}
           RUN_NUMBER: ${{ github.run_number }}
           RUN_ATTEMPT: ${{ github.run_attempt }}
+          KMP: true
