+++ .run/[KMP] Package Search Plugin.run.xml
@@ -1,6 +1,12 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Build Plugin" type="GradleRunConfiguration" factoryName="Gradle">
+  <configuration default="false" name="[KMP] Package Search Plugin" type="GradleRunConfiguration" factoryName="Gradle">
+    <log_file alias="idea.log" path="$PROJECT_DIR$/plugin/build/idea-sandbox/system/log/idea.log"/>
     <ExternalSystemSettings>
+      <option name="env">
+        <map>
+          <entry key="KMP" value="true"/>
+        </map>
+      </option>
       <option name="executionName" />
       <option name="externalProjectPath" value="$PROJECT_DIR$" />
       <option name="externalSystemIdString" value="GRADLE" />
@@ -10,7 +16,7 @@
       </option>
       <option name="taskNames">
         <list>
-          <option value=":plugin:buildShadowPlugin" />
+          <option value=":plugin:runIde"/>
         </list>
       </option>
       <option name="vmOptions" />
