+++ plugin/core/build.gradle.kts
@@ -43,19 +43,14 @@ kotlin.sourceSets.main {
 
 val pkgsPluginId: String by project
 
-val runNumber = System.getenv("RUN_NUMBER")?.toInt() ?: 0
-val runAttempt = System.getenv("RUN_ATTEMPT")?.toInt() ?: 0
-val snapshotMinorVersion = max(0, runNumber + runAttempt - 1)
-val versionString = project.version.toString()
-
 tasks {
     withType<Test> {
         environment("DB_PATH", layout.buildDirectory.file("tests/cache.db").get().asFile.absolutePath)
     }
     val generatePluginDataSources by registering(GeneratePackageSearchObject::class) {
+        group = "generate"
         pluginId = pkgsPluginId
         outputDir = generatedDir
-        pluginVersion = versionString.replace("-SNAPSHOT", ".$snapshotMinorVersion")
         packageName = "com.jetbrains.packagesearch.plugin.core"
     }
     sourcesJar {
