--- plugin/gradle/kmp/src/main/kotlin/com/jetbrains/packagesearch/plugin/gradle/KMPAttributes.kt
@@ -1,20 +1,0 @@
-package com.jetbrains.packagesearch.plugin.gradle
-
-
-data class KmpAttributesGroups(
-    val displayName: String,
-    val aggregationKeyword: String = displayName,
-)
-
-val KMP_ATTRIBUTES_GROUPS = listOf(
-    KmpAttributesGroups("iOS"),
-    KmpAttributesGroups("macOS"),
-    KmpAttributesGroups("tvOS"),
-    KmpAttributesGroups("watchOS"),
-    KmpAttributesGroups("JS"),
-    KmpAttributesGroups("JVM"),
-    KmpAttributesGroups("Linux"),
-    KmpAttributesGroups("Android"),
-    KmpAttributesGroups("WASM"),
-    KmpAttributesGroups("Windows", "mingw"),
-)
